filter client-side instead
This is a stop-gap before most of this communication is done via
websocket instead, meaning it doesn't
have the sluggishness of HTTP calls.

Having this one channel for state data also means that it'll be natural to
have deletions/updates move through it and be what informs the frontend as well.